requirements.txt
transformers
torch
numpy
pandas
scikit-learn
accelerate
optuna
msoffcrypto-tool
openpyxl
scipy
matplotlib

